[refactor] Logger Change to Header File
