+++ TestShell/TestShell.vcxproj
@@ -128,7 +128,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="CmdLancher.cpp" />
-    <ClCompile Include="Logger.cpp">
+    <ClCompile Include="Logger.h">
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Release|x64'"> %(AdditionalOptions)</AdditionalOptions>
     </ClCompile>
     <ClCompile Include="main.cpp">
