+++ TestShell/TestShell.vcxproj.filters
@@ -1,78 +1,78 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-	<ItemGroup>
-		<Filter Include="소스 파일">
-			<UniqueIdentifier>{4FC737F1-C7A5-4376-A066-2A32D752A2FF}</UniqueIdentifier>
-			<Extensions>cpp;c;cc;cxx;c++;cppm;ixx;def;odl;idl;hpj;bat;asm;asmx</Extensions>
-		</Filter>
-		<Filter Include="헤더 파일">
-			<UniqueIdentifier>{93995380-89BD-4b04-88EB-625FBE52EBFB}</UniqueIdentifier>
-			<Extensions>h;hh;hpp;hxx;h++;hm;inl;inc;ipp;xsd</Extensions>
-		</Filter>
-		<Filter Include="리소스 파일">
-			<UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
-			<Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
-		</Filter>
-		<Filter Include="테스트 파일">
-			<UniqueIdentifier>{9fdbe066-a1a6-4e73-9544-25cf7a7219b0}</UniqueIdentifier>
-		</Filter>
-	</ItemGroup>
-	<ItemGroup>
-		<ClCompile Include="main.cpp">
-			<Filter>소스 파일</Filter>
-		</ClCompile>
-		<ClCompile Include="test_cmdLauncher.cpp">
-			<Filter>테스트 파일</Filter>
-		</ClCompile>
-		<ClCompile Include="test_scripts.cpp">
-			<Filter>테스트 파일</Filter>
-		</ClCompile>
-		<ClCompile Include="TestScripts.cpp">
-			<Filter>소스 파일</Filter>
-		</ClCompile>
-		<ClCompile Include="test_shell.cpp">
-			<Filter>테스트 파일</Filter>
-		</ClCompile>
-		<ClCompile Include="CmdLancher.cpp">
-			<Filter>소스 파일</Filter>
-		</ClCompile>
-		<ClCompile Include="$(MSBuildThisFileDirectory)..\..\lib\native\src\gtest\gtest-all.cc">
-			<Filter>소스 파일</Filter>
-		</ClCompile>
-		<ClCompile Include="$(MSBuildThisFileDirectory)..\..\lib\native\src\gmock\gmock-all.cc">
-			<Filter>소스 파일</Filter>
-		</ClCompile>
-		<ClCompile Include="Logger.cpp">
-			<Filter>소스 파일</Filter>
-		</ClCompile>
-		<ClCompile Include="test_logger.cpp">
-			<Filter>테스트 파일</Filter>
-		</ClCompile>
-		<ClCompile Include="Runner.cpp">
-			<Filter>소스 파일</Filter>
-		</ClCompile>
-		<ClCompile Include="TestRunner.cpp">
-			<Filter>테스트 파일</Filter>
-		</ClCompile>
-	</ItemGroup>
-	<ItemGroup>
-		<None Include="packages.config" />
-	</ItemGroup>
-	<ItemGroup>
-		<ClInclude Include="test.h">
-			<Filter>헤더 파일</Filter>
-		</ClInclude>
-		<ClInclude Include="TestScripts.h">
-			<Filter>헤더 파일</Filter>
-		</ClInclude>
-		<ClInclude Include="CmdLancher.h">
-			<Filter>헤더 파일</Filter>
-		</ClInclude>
-		<ClInclude Include="Shell.h">
-			<Filter>헤더 파일</Filter>
-		</ClInclude>
-		<ClInclude Include="Runner.h">
-			<Filter>헤더 파일</Filter>
-		</ClInclude>
-	</ItemGroup>
+  <ItemGroup>
+    <Filter Include="소스 파일">
+      <UniqueIdentifier>{4FC737F1-C7A5-4376-A066-2A32D752A2FF}</UniqueIdentifier>
+      <Extensions>cpp;c;cc;cxx;c++;cppm;ixx;def;odl;idl;hpj;bat;asm;asmx</Extensions>
+    </Filter>
+    <Filter Include="헤더 파일">
+      <UniqueIdentifier>{93995380-89BD-4b04-88EB-625FBE52EBFB}</UniqueIdentifier>
+      <Extensions>h;hh;hpp;hxx;h++;hm;inl;inc;ipp;xsd</Extensions>
+    </Filter>
+    <Filter Include="리소스 파일">
+      <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
+      <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
+    </Filter>
+    <Filter Include="테스트 파일">
+      <UniqueIdentifier>{9fdbe066-a1a6-4e73-9544-25cf7a7219b0}</UniqueIdentifier>
+    </Filter>
+  </ItemGroup>
+  <ItemGroup>
+    <ClCompile Include="main.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="test_cmdLauncher.cpp">
+      <Filter>테스트 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="test_scripts.cpp">
+      <Filter>테스트 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="TestScripts.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="test_shell.cpp">
+      <Filter>테스트 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="CmdLancher.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="$(MSBuildThisFileDirectory)..\..\lib\native\src\gtest\gtest-all.cc">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="$(MSBuildThisFileDirectory)..\..\lib\native\src\gmock\gmock-all.cc">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="test_logger.cpp">
+      <Filter>테스트 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="Runner.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="TestRunner.cpp">
+      <Filter>테스트 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="Logger.h">
+      <Filter>헤더 파일</Filter>
+    </ClCompile>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="test.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="TestScripts.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="CmdLancher.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="Shell.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="Runner.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+  </ItemGroup>
 </Project>
